Restructured classes in the ChessServer\Socket namespace
