+++ cli/ratchet/data.php
@@ -5,8 +5,8 @@ namespace ChessServer\Cli\Ratchet;
 use ChessServer\Command\CommandParser;
 use ChessServer\Command\Data\CommandContainer;
 use ChessServer\Command\Data\Db;
-use ChessServer\Socket\RatchetClientStorage;
-use ChessServer\Socket\RatchetDataWebSocket;
+use ChessServer\Socket\Ratchet\RatchetClientStorage;
+use ChessServer\Socket\Ratchet\RatchetDataWebSocket;
 use Dotenv\Dotenv;
 use Monolog\Logger;
 use Monolog\Handler\StreamHandler;
