+++ cli/workerman/data.php
@@ -5,8 +5,8 @@ namespace ChessServer\Cli\Workerman;
 use ChessServer\Command\CommandParser;
 use ChessServer\Command\Data\CommandContainer;
 use ChessServer\Command\Data\Db;
-use ChessServer\Socket\WorkermanClientStorage;
-use ChessServer\Socket\WorkermanDataWebSocket;
+use ChessServer\Socket\Workerman\WorkermanClientStorage;
+use ChessServer\Socket\Workerman\WorkermanDataWebSocket;
 use Dotenv\Dotenv;
 use Monolog\Logger;
 use Monolog\Handler\StreamHandler;
